refactor: rename cardMotion to card-motion
